+++ graphistry/plotter.py
@@ -64,6 +64,7 @@ class Plotter(object):
         # Metadata
         self._name = None
         self._description = None
+        self._style = None
         # Integrations
         self._bolt_driver = None
         self._tigergraph = None
@@ -84,6 +85,113 @@ class Plotter(object):
             return pretty(rep)
         else:
             return str(rep)
+
+    def addStyle(self, fg=None, bg=None, page=None, logo=None):
+        """Set general visual styles
+        
+        See .bind() and .settings(url_params={}) for additional styling options, and style() for another way to set the same attributes.
+
+        To facilitate reuse and replayable notebooks, the addStyle() call is chainable. Invocation does not effect the old style: it instead returns a new Plotter instance with the new styles added to the existing ones. Both the old and new styles can then be used for different graphs.
+
+        addStyle() will extend the existing style settings, while style() will replace any in the same group
+
+        :param fg: Dictionary {'blendMode': str} of any valid CSS blend mode
+        :type fg: dict.
+
+        :param bg: Nested dictionary of page background properties. {'color': str, 'gradient': {'kind': str, 'position': str, 'stops': list }, 'image': { 'url': str, 'width': int, 'height': int, 'blendMode': str }
+        :type bg: dict.
+
+        :param logo: Nested dictionary of logo properties. { 'url': str, 'autoInvert': bool, 'position': str, 'dimensions': { 'maxWidth': int, 'maxHeight': int }, 'crop': { 'top': int, 'left': int, 'bottom': int, 'right': int }, 'padding': { 'top': int, 'left': int, 'bottom': int, 'right': int}, 'style': str}        
+        :type logo: dict.
+
+        :param page: Dictionary of page metadata settings. { 'favicon': str, 'title': str } 
+        :type page: dict.
+
+        :returns: Plotter.
+        :rtype: Plotter.
+
+        **Example: Chained merge - results in color, blendMode, and url being set**
+            ::
+                g2 =  g.addStyle(bg={'color': 'black'}, fg={'blendMode': 'screen'})
+                g3 = g2.addStyle(bg={'image': {'url': 'http://site.com/watermark.png'}})
+                
+        **Example: Overwrite - results in blendMode multiply**
+            ::
+                g2 =  g.addStyle(fg={'blendMode': 'screen'})
+                g3 = g2.addStyle(fg={'blendMode': 'multiply'})
+
+        **Example: Gradient background**
+            ::
+              g.addStyle(bg={'gradient': {'kind': 'linear', 'position': 45, 'stops': [['rgb(0,0,0)', '0%'], ['rgb(255,255,255)', '100%']]}})
+              
+        **Example: Page settings**
+            ::
+              g.addStyle(page={'title': 'Site - {{ name }}', 'favicon': 'http://site.com/logo.ico'})
+
+        """
+        style = copy.deepcopy(self._style or {})
+        o = {'fg': fg, 'bg': bg, 'page': page, 'logo': logo}
+        for k, v in o.items():
+            if not (v is None):
+                if isinstance(v, dict):
+                    if not (k in style) or (style[k] is None):
+                        style[k] = {}
+                    for k2, v2 in v.items():
+                        style[k][k2] = v2
+                else:
+                    style[k] = v
+        res = self.bind()
+        res._style = style
+        return res
+        
+
+
+    def style(self, fg=None, bg=None, page=None, logo=None):
+        """Set general visual styles
+        
+        See .bind() and .settings(url_params={}) for additional styling options, and addStyle() for another way to set the same attributes.
+
+        To facilitate reuse and replayable notebooks, the style() call is chainable. Invocation does not effect the old style: it instead returns a new Plotter instance with the new styles added to the existing ones. Both the old and new styles can then be used for different graphs.
+
+        style() will fully replace any defined parameter in the existing style settings, while addStyle() will merge over previous values
+
+        :param fg: Dictionary {'blendMode': str} of any valid CSS blend mode
+        :type fg: dict.
+
+        :param bg: Nested dictionary of page background properties. {'color': str, 'gradient': {'kind': str, 'position': str, 'stops': list }, 'image': { 'url': str, 'width': int, 'height': int, 'blendMode': str }
+        :type bg: dict.
+
+        :param logo: Nested dictionary of logo properties. { 'url': str, 'autoInvert': bool, 'position': str, 'dimensions': { 'maxWidth': int, 'maxHeight': int }, 'crop': { 'top': int, 'left': int, 'bottom': int, 'right': int }, 'padding': { 'top': int, 'left': int, 'bottom': int, 'right': int}, 'style': str}        
+        :type logo: dict.
+
+        :param page: Dictionary of page metadata settings. { 'favicon': str, 'title': str } 
+        :type page: dict.
+
+        :returns: Plotter.
+        :rtype: Plotter.
+
+        **Example: Chained merge - results in url and blendMode being set, while color is dropped**
+            ::
+                g2 =  g.style(bg={'color': 'black'}, fg={'blendMode': 'screen'})
+                g3 = g2.style(bg={'image': {'url': 'http://site.com/watermark.png'}})
+                
+        **Example: Gradient background**
+            ::
+              g.style(bg={'gradient': {'kind': 'linear', 'position': 45, 'stops': [['rgb(0,0,0)', '0%'], ['rgb(255,255,255)', '100%']]}})
+              
+        **Example: Page settings**
+            ::
+              g.style(page={'title': 'Site - {{ name }}', 'favicon': 'http://site.com/logo.ico'})
+
+        """        
+        style = copy.deepcopy(self._style or {})
+        o = {'fg': fg, 'bg': bg, 'page': page, 'logo': logo}
+        for k, v in o.items():
+            if not (v is None):
+                style[k] = v
+        res = self.bind()
+        res._style = style
+        return res
 
 
     def bind(self, source=None, destination=None, node=None,
@@ -395,18 +503,18 @@ class Plotter(object):
 
         api_version = PyGraphistry.api_version()
         if api_version == 1:
-            dataset = self._plot_dispatch(g, n, name, description, 'json')
+            dataset = self._plot_dispatch(g, n, name, description, 'json', self._style)
             if skip_upload:
                 return dataset
             info = PyGraphistry._etl1(dataset)
         elif api_version == 2:
-            dataset = self._plot_dispatch(g, n, name, description, 'vgraph')
+            dataset = self._plot_dispatch(g, n, name, description, 'vgraph', self._style)
             if skip_upload:
                 return dataset
             info = PyGraphistry._etl2(dataset)
         elif api_version == 3:
             PyGraphistry.refresh()
-            dataset = self._plot_dispatch(g, n, name, description, 'arrow')
+            dataset = self._plot_dispatch(g, n, name, description, 'arrow', self._style)
             if skip_upload:
                 return dataset
             #fresh
@@ -554,18 +662,18 @@ class Plotter(object):
                 util.error('%s attribute "%s" bound to "%s" does not exist.' % (typ, a, b))
 
 
-    def _plot_dispatch(self, graph, nodes, name, description, mode='json'):
+    def _plot_dispatch(self, graph, nodes, name, description, mode='json', metadata=None):
 
         if isinstance(graph, pandas.core.frame.DataFrame) \
             or isinstance(graph, pa.Table) \
             or ( not (maybe_cudf is None) and isinstance(graph, maybe_cudf.DataFrame) ):
-            return self._make_dataset(graph, nodes, name, description, mode)
+            return self._make_dataset(graph, nodes, name, description, mode, metadata)
 
         try:
             import igraph
             if isinstance(graph, igraph.Graph):
                 (e, n) = self.igraph2pandas(graph)
-                return self._make_dataset(e, n, name, description, mode)
+                return self._make_dataset(e, n, name, description, mode, metadata)
         except ImportError:
             pass
 
@@ -576,7 +684,7 @@ class Plotter(object):
                isinstance(graph, networkx.classes.multigraph.MultiGraph) or \
                isinstance(graph, networkx.classes.multidigraph.MultiDiGraph):
                 (e, n) = self.networkx2pandas(graph)
-                return self._make_dataset(e, n, name, description, mode)
+                return self._make_dataset(e, n, name, description, mode, metadata)
         except ImportError:
             pass
 
@@ -745,7 +853,7 @@ class Plotter(object):
         raise Exception('Unknown type %s: Could not convert data to Arrow' % str(type(table)))
 
 
-    def _make_dataset(self, edges, nodes, name, description, mode):
+    def _make_dataset(self, edges, nodes, name, description, mode, metadata=None):
         try:
             if len(edges) == 0:
                 util.warn('Graph has no edges, may have rendering issues')
@@ -753,17 +861,23 @@ class Plotter(object):
             1
 
         if mode == 'json':
+            if not (metadata is None):
+                if ('bg' in metadata) or ('fg' in metadata) or ('logo' in metadata) or ('page' in metadata):
+                    raise ValueError('Cannot set bg/fg/logo/page in api=1; try using api=3')
             edges_df = self._table_to_pandas(edges)
             nodes_df = self._table_to_pandas(nodes)
             return self._make_json_dataset(edges_df, nodes_df, name)
         elif mode == 'vgraph':
+            if not (metadata is None):
+                if ('bg' in metadata) or ('fg' in metadata) or ('logo' in metadata) or ('page' in metadata):
+                    raise ValueError('Cannot set bg/fg/logo/page in api=2; try using api=3')
             edges_df = self._table_to_pandas(edges)
             nodes_df = self._table_to_pandas(nodes)
             return self._make_vgraph_dataset(edges_df, nodes_df, name)
         elif mode == 'arrow':
             edges_arr = self._table_to_arrow(edges)
             nodes_arr = self._table_to_arrow(nodes)
-            return self._make_arrow_dataset(edges_arr, nodes_arr, name, description)
+            return self._make_arrow_dataset(edges=edges_arr, nodes=nodes_arr, name=name, description=description, metadata=metadata)
             #token=None, dataset_id=None, url_params = None)
         else:
             raise ValueError('Unknown mode: ' + mode)
@@ -809,7 +923,7 @@ class Plotter(object):
         dataset['encodings'] = encodings
         return dataset
 
-    def _make_arrow_dataset(self, edges: pa.Table, nodes: pa.Table, name: str, description: str) -> ArrowUploader:
+    def _make_arrow_dataset(self, edges: pa.Table, nodes: pa.Table, name: str, description: str, metadata) -> ArrowUploader:
         au = ArrowUploader(
             server_base_path=PyGraphistry.protocol() + '://' + PyGraphistry.server(),
             edges=edges, nodes=nodes,
@@ -820,6 +934,7 @@ class Plotter(object):
                 'agent': 'pygraphistry',
                 'apiversion' : '3',
                 'agentversion': sys.modules['graphistry'].__version__,
+                'metadata': metadata or {}
             },
             certificate_validation=PyGraphistry.certificate_validation())
         au.edge_encodings = au.g_to_edge_encodings(self)